# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RESQNET",
    "description": "Sprint 2: start-gated accessible allocation foundation",
    "version": "0.2.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Api Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios": {
      "get": {
        "summary": "Scenarios",
        "operationId": "scenarios_api_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioSummary"
                  },
                  "type": "array",
                  "title": "Response Scenarios Api Scenarios Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation": {
      "get": {
        "summary": "State",
        "operationId": "state_api_simulation_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationState"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/reset": {
      "post": {
        "summary": "Reset",
        "operationId": "reset_api_simulation_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/start": {
      "post": {
        "summary": "Start",
        "operationId": "start_api_simulation_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AllocationRequest": {
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AllocationRequest"
      },
      "AmbulanceData": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type_id": {
            "type": "string",
            "title": "Type Id"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "home_node_id": {
            "type": "string",
            "title": "Home Node Id"
          },
          "vehicle_class": {
            "type": "string",
            "title": "Vehicle Class"
          },
          "length_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Length M"
          },
          "width_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Width M"
          },
          "height_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Height M"
          },
          "approved_stretcher_positions": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Approved Stretcher Positions"
          },
          "approved_seated_positions": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Approved Seated Positions"
          },
          "occupied_stretcher_positions": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Occupied Stretcher Positions"
          },
          "occupied_seated_positions": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Occupied Seated Positions"
          },
          "equipment": {
            "additionalProperties": {
              "type": "integer",
              "minimum": 0.0
            },
            "type": "object",
            "title": "Equipment"
          },
          "crew": {
            "items": {
              "$ref": "#/components/schemas/CrewMember"
            },
            "type": "array",
            "title": "Crew"
          },
          "declared_capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Declared Capabilities"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Status"
          },
          "configuration_source": {
            "type": "string",
            "minLength": 1,
            "title": "Configuration Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "type_id",
          "node_id",
          "home_node_id",
          "vehicle_class",
          "length_m",
          "width_m",
          "height_m",
          "approved_stretcher_positions",
          "approved_seated_positions",
          "occupied_stretcher_positions",
          "occupied_seated_positions",
          "equipment",
          "crew",
          "declared_capabilities",
          "status",
          "configuration_source"
        ],
        "title": "AmbulanceData"
      },
      "CatalogEntry": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "description"
        ],
        "title": "CatalogEntry"
      },
      "CrewMember": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "role",
          "available"
        ],
        "title": "CrewMember"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hazard": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "flood",
              "debris",
              "gas"
            ],
            "title": "Kind"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "polygon": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "minItems": 3,
            "title": "Polygon"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "label",
          "active",
          "polygon",
          "note"
        ],
        "title": "Hazard"
      },
      "HospitalData": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "government",
              "authorized_private",
              "non_authorized_private"
            ],
            "title": "Category"
          },
          "authorization_reference": {
            "type": "string",
            "title": "Authorization Reference"
          },
          "beds_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Beds Total"
          },
          "beds_available": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Beds Available"
          },
          "icu_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Icu Total"
          },
          "icu_available": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Icu Available"
          },
          "facilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Facilities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "node_id",
          "category",
          "authorization_reference",
          "beds_total",
          "beds_available",
          "icu_total",
          "icu_available",
          "facilities"
        ],
        "title": "HospitalData"
      },
      "Node": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "x",
          "y"
        ],
        "title": "Node"
      },
      "PatientData": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "severity": {
            "type": "string",
            "enum": [
              "critical",
              "urgent",
              "stable"
            ],
            "title": "Severity"
          },
          "transport_position": {
            "type": "string",
            "enum": [
              "stretcher",
              "seated"
            ],
            "title": "Transport Position"
          },
          "required_capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Capabilities"
          },
          "required_equipment": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Equipment"
          },
          "required_crew_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Crew Roles"
          },
          "required_facilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Facilities"
          },
          "clinical_compatibility_group": {
            "type": "string",
            "title": "Clinical Compatibility Group"
          },
          "status": {
            "type": "string",
            "const": "waiting",
            "title": "Status"
          },
          "required_hospital_place": {
            "type": "string",
            "enum": [
              "general",
              "icu"
            ],
            "title": "Required Hospital Place",
            "default": "general"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "incident_id",
          "node_id",
          "severity",
          "transport_position",
          "required_capabilities",
          "required_equipment",
          "required_crew_roles",
          "required_facilities",
          "clinical_compatibility_group",
          "status"
        ],
        "title": "PatientData"
      },
      "ResetRequest": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "enum": [
              "flood",
              "building_collapse",
              "industrial_gas_leak"
            ],
            "title": "Scenario Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_id"
        ],
        "title": "ResetRequest"
      },
      "RoadData": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "distance_km": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Distance Km"
          },
          "travel_time_min": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Travel Time Min"
          },
          "max_width_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Width M"
          },
          "max_height_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Height M"
          },
          "allowed_vehicle_classes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Allowed Vehicle Classes"
          },
          "blocked": {
            "type": "boolean",
            "title": "Blocked"
          },
          "hazard_zone_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hazard Zone Id"
          },
          "restriction_note": {
            "type": "string",
            "title": "Restriction Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source",
          "target",
          "distance_km",
          "travel_time_min",
          "max_width_m",
          "max_height_m",
          "allowed_vehicle_classes",
          "blocked",
          "hazard_zone_id",
          "restriction_note"
        ],
        "title": "RoadData"
      },
      "ScenarioSummary": {
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "flood",
              "building_collapse",
              "industrial_gas_leak"
            ],
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description"
        ],
        "title": "ScenarioSummary"
      },
      "SimulationState": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version"
          },
          "scenario_id": {
            "type": "string",
            "enum": [
              "flood",
              "building_collapse",
              "industrial_gas_leak"
            ],
            "title": "Scenario Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/Node"
            },
            "type": "array",
            "minItems": 1,
            "title": "Nodes"
          },
          "hazard_zones": {
            "items": {
              "$ref": "#/components/schemas/Hazard"
            },
            "type": "array",
            "title": "Hazard Zones"
          },
          "roads": {
            "items": {
              "$ref": "#/components/schemas/RoadData"
            },
            "type": "array",
            "minItems": 1,
            "title": "Roads"
          },
          "ambulance_types": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            },
            "type": "array",
            "title": "Ambulance Types"
          },
          "equipment_types": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            },
            "type": "array",
            "title": "Equipment Types"
          },
          "ambulances": {
            "items": {
              "$ref": "#/components/schemas/AmbulanceData"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ambulances"
          },
          "hospitals": {
            "items": {
              "$ref": "#/components/schemas/HospitalData"
            },
            "type": "array",
            "minItems": 1,
            "title": "Hospitals"
          },
          "patients": {
            "items": {
              "$ref": "#/components/schemas/PatientData"
            },
            "type": "array",
            "minItems": 1,
            "title": "Patients"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "synthetic": {
            "type": "boolean",
            "const": true,
            "title": "Synthetic",
            "default": true
          },
          "run_status": {
            "type": "string",
            "enum": [
              "idle",
              "allocated"
            ],
            "title": "Run Status",
            "default": "idle"
          },
          "allocations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Allocations"
          },
          "reservations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Reservations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "scenario_id",
          "title",
          "description",
          "nodes",
          "hazard_zones",
          "roads",
          "ambulance_types",
          "equipment_types",
          "ambulances",
          "hospitals",
          "patients",
          "revision"
        ],
        "title": "SimulationState"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}